pyproject.toml
[project]
name = "giga-trader"
version = "1.0.0"
description = "ML-based SPY swing trading system"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = {text = "MIT"}

dependencies = [
    # Core Data Science
    "numpy>=1.26.0,<2.0.0",
    "pandas>=2.1.0",
    "scipy>=1.11.0",

    # Machine Learning
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "imbalanced-learn>=0.11.0",

    # Market Data Sources
    "alpaca-py>=0.21.0",
    "yfinance>=0.2.30",

    # Technical Analysis
    "ta>=0.11.0",

    # Statistics
    "statsmodels>=0.14.0",

    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",

    # Configuration & Utils
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.5.0",
    "typer>=0.9.0",

    # Database & Storage
    "sqlalchemy>=2.0.0",
    "pyarrow>=14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
